// todo-webapp/src/generated/todo-api.ts
export interface components {
    schemas: {
        TodoEntry: {
            /** @description entry id */
            id: string;
            /** @description entry title */
            title: string;
            /** @description completion state */
            done: boolean;
            /** Format: date-time */
            createdAt: string;
            /** Format: date-time */
            updatedAt: string;
        };
        TodoEntryCreate: {
            /** @description entry title */
            title: string;
        };
        TodoEntryUpdate: {
            /** @description new title */
            title?: string;
            /** @description new completion state */
            done?: boolean;
        };
        Error: {
            /** @description HTTP or application error code */
            code: number;
            /** @description short human-readable label */
            message: string;
            /** @description detailed explanation */
            description?: string;
            /** @description URI to documentation */
            moreInfo?: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}